starCount와 constellationCount를 int형으로 변경

See merge request s09-webmobile2-sub2/S09P12A410!134

+++ unibirth-BE/src/main/java/com/ssafy/unibirth/constellation/service/ConstellationService.java
@@ -34,7 +34,7 @@ public class ConstellationService {
 
     public CreateConstellationResDto create(ConstellationReqDto dto) {
         Member member = memberService.getCurrentMember();
-        Long constellationLimit = member.minusConstellationLimit();
+        int constellationLimit = member.minusConstellationLimit();
 
         Planet planet = planetService.findPlanetById(dto.getPlanetId());
         List<List<Integer>> pointList = insertZPoint(dto.getPointList());
